pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chesscom-game-fetcher"
version = "0.0.1"
authors = [
  { name="Scott Oxtoby", email="pypi@oxtoby.uk" },
]
description = "Fetches all games from chess.com with rating specifiers"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "chess.com",
    "requests",
    "python-dateutil",
    "prettytable"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
